Add -Wall -Wextra, fix warnings
git-svn-id: https://svn.code.sf.net/p/apertium/svn/trunk/apertium-lex-tools@66150 72bbbca6-d526-0410-a7d9-f06f51895060

+++ scripts/process_tagger_output.cc
@@ -15,7 +15,7 @@
 using namespace std;
 
 int find(vector<wstring> xs, wstring x) {
-	for (int i = 0; i < xs.size(); i++) {
+	for (size_t i = 0; i < xs.size(); i++) {
 		if (xs[i] == x)
 			return i;
 	}
@@ -37,7 +37,7 @@ vector<wstring> parseTags(wstring token) {
 	int state = 0; // outside
 	vector<wstring> tags;
 	wstring buffer;
-	for(int i = 0; i < token.size(); i++) {
+	for (size_t i = 0; i < token.size(); i++) {
 		wchar_t c = token[i];
 		if (state == 0) {
 			if (c == '<') {
@@ -60,7 +60,7 @@ vector<wstring> wsplit(wstring wstr, wchar_t delim) {
 	vector<wstring> tokens;
 	wstring buffer;
 
-	for(int i = 0; i < wstr.size(); i++) {
+	for (size_t i = 0; i < wstr.size(); i++) {
 		buffer += wstr[i];
 		if(wstr[i] == delim) {
 			tokens.push_back(buffer);
@@ -76,7 +76,7 @@ vector<wstring> wsplit(wstring wstr, wchar_t delim) {
 
 wstring getLemma(wstring token) {
 	wstring buffer;
-	for(int i = 0; i < token.size(); i++) {
+	for (size_t i = 0; i < token.size(); i++) {
 		if(token[i] != '<') {
 			buffer += token[i];
 		} else {
@@ -117,7 +117,7 @@ void processTaggerOutput(FSTProcessor *bilingual) {
 				vector<wstring> trimmedTags = parseTags(targetTrimmed);
 				vector<wstring> newTags;
 
-				for (int i = 0; i < sourceTags.size(); i++) {
+				for (size_t i = 0; i < sourceTags.size(); i++) {
 					wstring sourceTag = sourceTags[i];
 					int idx_1 = find(targetTags, sourceTag);
 					int idx_2 = find(trimmedTags, sourceTag);
@@ -126,7 +126,7 @@ void processTaggerOutput(FSTProcessor *bilingual) {
 					}
 				}
 				wcout << getLemma(buffer);
-				for(int i = 0; i < newTags.size(); i++) {
+				for (size_t i = 0; i < newTags.size(); i++) {
 					wcout << '<' << newTags[i] << '>';
 				}
 				targetTrimmed[0] = '/';
@@ -135,7 +135,7 @@ void processTaggerOutput(FSTProcessor *bilingual) {
 					wcout << L"/" + buffer + L"$";
 				} else {
 					vector<wstring> tokens = wsplit(targetTrimmed, '/');
-					for(int i = 0; i < tokens.size(); i++) {
+					for (size_t i = 0; i < tokens.size(); i++) {
 						wcout << tokens[i];
 					}
 				}
